customs-analytics-v2/backend/requirements.txt
# Customs Revenue & Trade Volume Analytics - Version 2
# Python >= 3.11 Required

# Core Framework
fastapi==0.110.*
uvicorn[standard]==0.29.*
pydantic==2.6.*
pydantic-settings==2.2.*

# Database (SQLAlchemy Core)
sqlalchemy==2.0.*
asyncpg==0.29.*
alembic==1.13.*

# Analytics & Forecasting
pandas==2.2.*
numpy==1.26.*
scikit-learn==1.4.*
statsmodels==0.14.*

# Testing & Utilities
pytest==8.1.*
pytest-asyncio==0.23.*
httpx==0.27.*
python-dotenv==1.0.*
